usm: http2: Save last frame header while filtering
In a followup commit, we will need to have the last frame header for the case of incomplete frame payload.
For that matter, we save the last frame header from the function 'pktbuf_find_relevant_frames'.

In case we do have an incomplete frame payload, we need to know whether the frame is interesting (for further
processing) and how many bytes we need to read from future packets.