usm: http2: Mark some frames as processed
Added a 'processed' flag to the frame_payload struct. We want to avoid from processing some frames multiple times, and
the flag is meant to ensure it.

For data frames we only process the header and we don't care about their payloads, so we need to ensure
we processed the frame header only once. But for header frames we do process the payload, and we still need
to do so as a best effort. A common practice is to send only the frame header for Headers frame, and in a
separate packet to send the payload. In such a case, we will have a payload remainder, and we want to process it