omnibus: otel-agent: remove platform assertion
Such conditionals in the project will cause the exception to be raised
when the project is loaded, even if it isn't meant to be built.
Even moving it to the software without additional condition will cause
the exception to trigger.
Since this check doesn't serve much purpose aside from being self
documenting, we can remove it in favor of passing builds